refactor: DRY filter code, misc style tweaks
