+++ src/components/search/Filters/Chips/SectionTypeFilterChip.tsx
@@ -1,3 +1,5 @@
+import type { FilterBarChipProps } from '@/modules/filters';
+import { displaySectionTypeName } from '@/modules/semesters';
 import {
   Checkbox,
   Divider,
@@ -6,36 +8,13 @@ import {
   MenuList,
   Tooltip,
 } from '@mui/material';
-import React, { type Dispatch, type SetStateAction } from 'react';
 import FilterChip from '../FilterChip';
-import { type FilterBarChipProps } from '../utils';
 
 type SectionTypeFilterChipProps = FilterBarChipProps<{
   sectionTypes: string[];
   chosenSectionTypes: string[];
-  setChosenSectionTypes: Dispatch<SetStateAction<string[]>>;
+  setChosenSectionTypes: React.Dispatch<React.SetStateAction<string[]>>;
 }>;
-
-function displaySectionTypeName(id: string): string {
-  const SectionTypesMap: Record<string, string> = {
-    '0xx': 'Normal day lecture',
-    '0Wx': 'Online class',
-    '0Hx': 'Hybrid day class (online + face-to-face)',
-    '0Lx': 'LLC-only section',
-    '5Hx': 'Hybrid night class (online + face-to-face)',
-    '1xx': 'Lab section (sciences)',
-    '2xx': 'Discussion section (humanities)',
-    '3xx': 'Problem section (maths)',
-    '5xx': 'Night lecture (past 5 PM)',
-    '6xx': 'Lab night section (past 7 PM)',
-    '7xx': 'Exam section',
-    HNx: 'Honors-only',
-    HON: 'Honors-only',
-    xUx: 'Summer Class',
-  };
-
-  return SectionTypesMap[id] || id; // Default to ID if no mapping exists
-}
 
 export default function SectionTypeFilterChip({
   type,
