+++ src/modules/semesters.tsx
@@ -1,23 +1,8 @@
 import type { SearchResult } from '@/types/SearchQuery';
 
-const typeRegexes: Record<string, RegExp> = {
-  '0xx': /0[0-9][0-9]/,
-  '0Wx': /0W[0-9]/,
-  '0Hx': /0H[0-9]/,
-  '0Lx': /0L[0-9]/,
-  '5Hx': /5H[0-9]/,
-  '1xx': /1[0-9][0-9]/,
-  '2xx': /2[0-9][0-9]/,
-  '3xx': /3[0-9][0-9]/,
-  '5xx': /5[0-9][0-9]/,
-  '6xx': /6[0-9][0-9]/,
-  '7xx': /7[0-9][0-9]/,
-  HNx: /HN[0-9]/,
-  HON: /HON/,
-  xUx: /[0-9]U[0-9]/,
-  OTHERS: /.*/,
-};
-const others = 'OTHERS';
+////////////////////////////////////////////////////////////////////////////////
+// SEMESTERS
+////////////////////////////////////////////////////////////////////////////////
 
 /** A comparator function used when sorting semesters by name. Negative if semester `a` is older than `b`, positive if newer, zero if equal. */
 export function compareSemesters(a: string, b: string): number {
@@ -37,6 +22,44 @@ export function getCurrentSemester(): string {
   const month = today.getMonth();
   const semester = month < 5 ? 'S' : month < 8 ? 'U' : 'F';
   return `${year}${semester}`;
+}
+
+/**
+ * Returns long-semesters (i.e. not summer semesters) offered recently. By default, returns long-semesters from the last 2 years
+ * @param {string[]} available Array of available semesters
+ * @param {number} [maximumNumber=4] Maximum number of recent semesters to return (default 4)
+ */
+export function getRecentSemesters(
+  available: string[],
+  maximumNumber: number = 4, // By default, return up to the last 4 long-semesters
+) {
+  // get current month and year
+  const today = new Date();
+  const mm = today.getMonth() + 1; // January is 1
+  let yyyy = today.getFullYear();
+
+  let season = 'F';
+  if (mm <= 5)
+    // jan - may
+    season = 'S';
+  else season = 'F';
+
+  // generate recent semesters dynamically from the current day
+  const recentSemesters: string[] = [];
+  for (let i = maximumNumber; i >= 1; i--) {
+    if (season === 'S') {
+      // then the previous semester is last year's Fall
+      yyyy = yyyy - 1;
+      season = 'F';
+    } else {
+      // then the previous long-semester is this year's Spring
+      season = 'S';
+    }
+
+    recentSemesters.push(yyyy.toString().substring(2) + season);
+  }
+
+  return recentSemesters.filter((value) => available.includes(value));
 }
 
 /** Semesters in `available` that fall in the inclusive window between `current` and `latest`. If `latest` is empty, keeps terms >= `current`.*/
@@ -83,11 +106,58 @@ export function getSemestersFromSearchResults(searchResults: SearchResult[]) {
     ...new Set(searchResults.flatMap((r) => r.grades.map((g) => g._id))),
   ].sort((a, b) => compareSemesters(b, a));
 }
+
 export function getSemestersFromSearchResult(searchResult: SearchResult) {
   return [...new Set(searchResult.grades.map((g) => g._id))].sort((a, b) =>
     compareSemesters(b, a),
   );
 }
+
+////////////////////////////////////////////////////////////////////////////////
+// SECTIONS
+////////////////////////////////////////////////////////////////////////////////
+
+const typeRegexes: Record<string, RegExp> = {
+  '0xx': /0[0-9][0-9]/,
+  '0Wx': /0W[0-9]/,
+  '0Hx': /0H[0-9]/,
+  '0Lx': /0L[0-9]/,
+  '5Hx': /5H[0-9]/,
+  '1xx': /1[0-9][0-9]/,
+  '2xx': /2[0-9][0-9]/,
+  '3xx': /3[0-9][0-9]/,
+  '5xx': /5[0-9][0-9]/,
+  '6xx': /6[0-9][0-9]/,
+  '7xx': /7[0-9][0-9]/,
+  HNx: /HN[0-9]/,
+  HON: /HON/,
+  xUx: /[0-9]U[0-9]/,
+  OTHERS: /.*/,
+};
+const others = 'OTHERS';
+
+export function displaySectionTypeName(id: string): string {
+  const SectionTypesMap: Record<string, string> = {
+    '0xx': 'Normal day lecture',
+    '0Wx': 'Online class',
+    '0Hx': 'Hybrid day class (online + face-to-face)',
+    '0Lx': 'LLC-only section',
+    '5Hx': 'Hybrid night class (online + face-to-face)',
+    '1xx': 'Lab section (sciences)',
+    '2xx': 'Discussion section (humanities)',
+    '3xx': 'Problem section (maths)',
+    '5xx': 'Night lecture (past 5 PM)',
+    '6xx': 'Lab night section (past 7 PM)',
+    '7xx': 'Exam section',
+    HNx: 'Honors-only',
+    HON: 'Honors-only',
+    xUx: 'Summer Class',
+    OTHERS: 'Others',
+  };
+
+  return SectionTypesMap[id] || id; // Default to ID if no mapping exists
+}
+
 export function getSectionTypesFromSearchResults(
   searchResults: SearchResult[],
 ) {
@@ -99,6 +169,7 @@ export function getSectionTypesFromSearchResults(
     ),
   ].sort();
 }
+
 export function getSectionTypesFromSearchResult(searchResult: SearchResult) {
   return [
     ...new Set(
